+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/book/controller/BookController.java
@@ -42,14 +42,14 @@ import sigma.chackcheck.domain.user.service.UserService;
 
 @Controller
 @RequiredArgsConstructor
-@RequestMapping("/api")
+@RequestMapping("/api/books")
 public class BookController {
 
     private final BookService bookService;
     private final BookBorrowService bookBorrowService;
     private final UserService userService;
 
-    @GetMapping("/books/all")
+    @GetMapping("/all")
     public ApiResponse<SuccessBody<BookPageResponse>> getMainPage(
         @RequestParam(value = "page", defaultValue = "0") int page) {
 
@@ -58,7 +58,7 @@ public class BookController {
         return getBookSuccessBodyApiResponse(page, bookList);
     }
 
-    @GetMapping("/books/category")
+    @GetMapping("/category")
     public ApiResponse<SuccessBody<BookPageResponse>> getBooksByCategoryName(
         @RequestParam(value = "categoryName") String categoryName,
         @RequestParam(value = "page", defaultValue = "0") int page
@@ -68,7 +68,7 @@ public class BookController {
         return getBookSuccessBodyApiResponse(page, bookList);
     }
 
-    @GetMapping("/books/all/search")
+    @GetMapping("/all/search")
     public ApiResponse<SuccessBody<BookPageResponse>> getBooksBySearch(
         @RequestParam(value = "keyword") String keyword,
         @RequestParam(value = "page", defaultValue = "0") int page
@@ -78,16 +78,7 @@ public class BookController {
         return getBookSuccessBodyApiResponse(page, bookList);
     }
 
-    @GetMapping("/admin/books/approve")
-    public ApiResponse<SuccessBody<BookApprovePageResponse>> getBookApprovePage(
-        @RequestParam(value = "page", defaultValue = "0") int page
-    ){
-        Page<BookApprove> bookApproveList = bookService.getBookApprovePage(page);
-
-        return getBookApproveSuccessBodyApiResponse(page, bookApproveList);
-    }
-
-    @GetMapping("/books/{bookId}")
+    @GetMapping("/{bookId}")
     public ApiResponse<SuccessBody<BookDetailPageResponse>> getBookDetails(
         @PathVariable(value = "bookId") Long bookId
     ){
@@ -119,15 +110,9 @@ public class BookController {
         return ApiResponseGenerator.success(bookDetailPageResponse, HttpStatus.OK, SuccessMessage.GET);
     }
 
-    @PostMapping("/books")
+    @PostMapping()
     public ApiResponse<SuccessBody<Void>> createBookApprove(@RequestBody CreateBookApproveRequest createBookApproveRequest){
         bookService.createBookApprove(createBookApproveRequest);
-        return ApiResponseGenerator.success(HttpStatus.CREATED, SuccessMessage.CREATE);
-    }
-
-    @PostMapping("/admin/books/approve")
-    public ApiResponse<SuccessBody<Void>> approveBookApprove(@RequestBody CreateBookRequest createBookRequest){
-        bookService.createBook(createBookRequest);
         return ApiResponseGenerator.success(HttpStatus.CREATED, SuccessMessage.CREATE);
     }
 
@@ -143,19 +128,5 @@ public class BookController {
         BookPageResponse bookPageResponse = BookPageResponse.of(pageInfo, bookDtoList);
 
         return ApiResponseGenerator.success(bookPageResponse, HttpStatus.OK, SuccessMessage.GET);
-    }
-
-    private ApiResponse<SuccessBody<BookApprovePageResponse>> getBookApproveSuccessBodyApiResponse(
-        @RequestParam(value = "page", defaultValue = "0") int page,
-        Page<BookApprove> bookList) {
-        PageInfo pageInfo =  PageInfo.of(page, bookList.getTotalElements(), bookList.getTotalPages());
-
-        List<BookApproveDTO> bookApproveDtoList = bookList.getContent().stream()
-            .map(bookApprove -> BookApproveDTO.of(bookApprove, userService.findById(bookApprove.getUserId())))
-            .toList();
-
-        BookApprovePageResponse bookApprovePageResponse = BookApprovePageResponse.of(pageInfo, bookApproveDtoList);
-
-        return ApiResponseGenerator.success(bookApprovePageResponse, HttpStatus.OK, SuccessMessage.GET);
     }
 }
